REFACTOR: change credit score return type

+++ src/main/java/com/creditscoreservice/application/creditscore/entity/CreditScore.java
@@ -16,6 +16,6 @@ public class CreditScore {
     private Long id;
 
    @Column(name = "creditScore", nullable = false)
-    private float creditScore;
+    private int creditScore;
 
 }
